+++ Changelog.md
@@ -1,4 +1,6 @@
 # Version 1.1.0
 - Added Onchopristis
+- Added a new music disc "Doomsurf" by ChipsTheCat
 - Fixed Lepidodendron Wood not being craftable
 - Added targeting tag for Carnotaurus
+- Added a seperate sound event for Megalania step sounds